@@ -16,12 +16,11 @@
 
 package com.cpur.data;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.annotation.NonNull;
 
 import java.util.List;
-import java.util.Map;
 
 
 public class StoryRepository {
