@@ -16,8 +16,8 @@
 
 package com.cpur.data.source.local;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
 
 import com.cpur.data.Paragraph;
 import com.cpur.data.Story;
