@@ -1,7 +1,7 @@
 package com.cpur.data;
 
-import android.arch.persistence.room.Embedded;
-import android.arch.persistence.room.Relation;
+import androidx.room.Embedded;
+import androidx.room.Relation;
 
 import java.util.ArrayList;
 import java.util.List;
