@@ -2,56 +2,46 @@ package com.cpur.notifications;
 
 import android.util.Log;
 
+import androidx.annotation.NonNull;
+
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.iid.FirebaseInstanceIdService;
+import com.google.firebase.messaging.FirebaseMessagingService;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class CpurFirebaseInstanceIDService extends FirebaseInstanceIdService {
+public class CpurFirebaseInstanceIDService extends FirebaseMessagingService {
 
-        private static final String TAG = "MyFirebaseIIDService";
+    private static final String TAG = "MyFirebaseIIDService";
 
-        /**
-         * Called if InstanceID token is updated. This may occur if the security of
-         * the previous token had been compromised. Note that this is called when the InstanceID token
-         * is initially generated so this is where you would retrieve the token.
-         */
-        // [START refresh_token]
-        @Override
-        public void onTokenRefresh() {
-            // Get updated InstanceID token.
-            String refreshedToken = FirebaseInstanceId.getInstance().getToken();
-            Log.d(TAG, "Refreshed token: " + refreshedToken);
+    @Override
+    public void onNewToken(@NonNull String s) {
+        super.onNewToken(s);
+        sendRegistrationToServer(s);
+    }
 
-            // If you want to send messages to this application instance or
-            // manage this apps subscriptions on the server side, send the
-            // Instance ID token to your app server.
-            sendRegistrationToServer(refreshedToken);
+
+    /**
+     * Persist token to third-party servers.
+     * <p>
+     * Modify this method to associate the user's FCM InstanceID token with any server-side account
+     * maintained by your application.
+     *
+     * @param token The new token.
+     */
+    private void sendRegistrationToServer(String token) {
+        String uid = FirebaseAuth.getInstance().getUid();
+        if (uid != null) {
+            DatabaseReference mDatabase = FirebaseDatabase.getInstance().getReference();
+            DatabaseReference databaseReference = mDatabase.child("users").child(uid);
+            Map<String, Object> stringObjectMap = new HashMap<>();
+            stringObjectMap.put("notificationToken", token);
+            databaseReference.updateChildren(stringObjectMap);
         }
-        // [END refresh_token]
 
-        /**
-         * Persist token to third-party servers.
-         *
-         * Modify this method to associate the user's FCM InstanceID token with any server-side account
-         * maintained by your application.
-         *
-         * @param token The new token.
-         */
-        private void sendRegistrationToServer(String token) {
-            String uid = FirebaseAuth.getInstance().getUid();
-            if (uid != null){
-                DatabaseReference mDatabase = FirebaseDatabase.getInstance().getReference();
-                DatabaseReference databaseReference = mDatabase.child("users").child(uid);
-                Map<String, Object> stringObjectMap = new HashMap<>();
-                stringObjectMap.put("notificationToken", token);
-                databaseReference.updateChildren(stringObjectMap);
-            }
-
-        }
+    }
 
 }
